more updates in javadocs and site documentation, also misc improvements in common hssf-xssf interfaces
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@714128 13f79535-47bb-0310-9956-ffa450edef68

+++ src/examples/src/org/apache/poi/xssf/usermodel/examples/NewLinesInCells.java
@@ -47,7 +47,7 @@ public class NewLinesInCells {
         row.setHeightInPoints((2*sheet.getDefaultRowHeightInPoints()));
 
         //adjust column width to fit the content
-        sheet.autoSizeColumn((short)2);
+        sheet.autoSizeColumn(2);
 
         FileOutputStream fileOut = new FileOutputStream("ooxml-newlines.xlsx");
         wb.write(fileOut);
